Normalize the depth of the LiA and LiD JSON structures
An artifact of how we convert XML to a PHP array and then to JSON is
that single-item LiA and LiD lists (that is, when there's just one
contribution or just one expense) are structured differently than
reports with multiple LiA or LiDs. That inconsistency makes it awkward
to handle the data, and doesn't accurately reflect the XML structure.
Here's a shim to solve that.